backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pymongo==4.6.1
pika==1.3.1
python-dotenv==1.0.1
pydantic==2.6.1
python-jose==3.3.0
python-multipart==0.0.9
passlib==1.7.4
bcrypt==4.1.2 